+++ com.unity.postprocessing/PostProcessing/Runtime/Effects/AutoExposure.cs
@@ -85,7 +85,7 @@ namespace UnityEngine.Rendering.PostProcessing
             return enabled.value
                 && SystemInfo.supportsComputeShaders
                 && !RuntimeUtilities.isAndroidOpenGL
-                && !RuntimeUtilities.isWebGL
+                && !RuntimeUtilities.isWebNonWebGPU
                 && RenderTextureFormat.RFloat.IsSupported()
                 && context.resources.computeShaders.autoExposure
                 && context.resources.computeShaders.exposureHistogram;