+++ com.unity.postprocessing/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Disabled compute based effects not supported on WebGL and Android OpenGL (IN-2999)
 - Fixed grid gizmo is visible through geometry when Post Process Layer is enabled (IN-10318)
 - Fixed transparent objects rendering incorrectly when TAA is disabled (IN-31494)
+- Fixed MSVO to support platforms with limited storage texture support.
+- added WebGPU API platform defines
 
 ## [3.3.0] - 2023-05-11